requirements.txt
python-telegram-bot==13.7
python-dotenv==0.19.0
requests==2.28.1
Flask==2.0.1